+++ src/utils/handlers/misc/mod.rs
@@ -4,3 +4,4 @@ pub mod joke_call;
 pub mod exceptions;
 pub mod everyone_case;
 pub mod attachment_case;
+pub mod link_spam_handler;
